fix(stella-pipeline): reconcile verdict-reuse merge — unbreak main
Four commits landed within ~2 minutes after 7c3ab928 and merged cleanly
textually but were semantically incompatible: bed6a87e (#1462, token-budgeted
diff rendering — DiffContext, delta framing via CandidateState::last_verdict_diff,
witness stat-lining, DIFF_STAT_LINE_NOTE) was authored before 6f8f5e8c (#1482,
verdict-reuse caching via CandidateState::last_verdict + witness hunk stripping
via strip_witness_hunks). 6f8f5e8c's diff was built against a stale pre-bed6a87e
verify.rs: it deleted DIFF_STAT_LINE_NOTE and diff_render::DiffContext wiring
that bed6a87e had just added (while the two prompt functions that still
interpolated {DIFF_STAT_LINE_NOTE} were untouched, orphaning the constant), it
reintroduced the old char-based bounded_worker_diff/VERIFIER_DIFF_BUDGET_CHARS
clamp that bed6a87e had deleted in favor of diff_render's per-segment renderer,
it renamed last_verdict_diff to last_verdict instead of adding a second field,
and it dropped the 7th `diff_ctx` argument from the `verifier()` call site
without updating the still-7-argument function signature.

Reconciled so both features' intent survives:
- CandidateState keeps both `last_verdict` (verdict-reuse cache, keyed by an
  inputs digest) and `last_verdict_diff` (delta-framing baseline for
  diff_render), since they serve different rounds of the same escalation.
- verify.rs keeps diff_render-based rendering (DIFF_STAT_LINE_NOTE restored)
  and drops the dead reintroduced char-clamp duplicate.
- The ModelVerdict arm now strips witness hunks, checks the reuse cache, and
  on a cache miss calls the verifier with a full DiffContext (witness_paths +
  the delta-framing baseline), matching both PRs' designs together.

Also fixed, uncovered only once the crate compiled again:
- Two rustdoc intra-doc links in stella-observatory's `execution_journal` doc
  comment pointed at private items (`recall_timings`, `JOURNAL_BODY_CLIP`);
  changed to plain code spans.
- Two `clippy::unnecessary_map_or` and three `clippy::too_many_arguments`
  lints that were latent in bed6a87e's own diff (masked by the compile
  errors, never actually clippy-checked before this fix) — fixed with
  `is_none_or` and `#[allow(clippy::too_many_arguments)]` plus a one-line
  justification each, since the arity is inherent to threading prompt
  shaping, leak scrubbing, and call metering through one escalation call.

Brought pipeline.rs (3851 -> ~3620 lines) and pipeline/tests.rs (2664 -> 2608
lines) back under the file-size ratchet by extracting the verdict-reuse/
witness-exclusion logic into `Pipeline::resolve_verdict` and
`Pipeline::distress_guidance_reason` (pipeline/verifier_stage.rs), the
evidence-summary assembly into `Pipeline::model_verdict_evidence_summary`
(pipeline/verify_probes.rs), and the `ScriptedProvider` test double into its
own file (pipeline/tests/scripted_provider.rs) — the same sibling-submodule
pattern as stella-core's driver/settlement.rs split. No baseline edit needed.

Build fix, no witness test: verified via `cargo test -p stella-pipeline`
(495 lib + all integration suites green, including
`an_unchanged_revision_reuses_the_verdict_instead_of_rebuying_it`),
`cargo clippy -p stella-pipeline -p stella-observatory --all-targets -- -D
warnings`, `RUSTDOCFLAGS="-D warnings" cargo doc -p stella-observatory
--no-deps`, and `./scripts/check-file-size.sh`, all exit 0.
