+++ crates/stella-pipeline/src/pipeline.rs
@@ -822,11 +822,20 @@ struct CandidateState {
     /// next revision is told (`witness::airlock`).
     failures: Vec<FailureFingerprint>,
     /// The last model verdict this candidate bought, keyed by a digest of the
-    /// exact inputs that produced it (#1431). A revision that changed nothing
-    /// the verdict depends on reuses the opinion instead of re-buying it —
-    /// the verifier is stateless across rounds, so identical inputs are the
-    /// same question, and paying twice buys only sampling noise.
+    /// exact inputs that produced it (#1431 step 1). A revision that changed
+    /// nothing the verdict depends on reuses the opinion instead of
+    /// re-buying it — the verifier is stateless across rounds, so identical
+    /// inputs are the same question, and paying twice buys only sampling
+    /// noise.
     last_verdict: Option<(u64, crate::verify::Verdict)>,
+    /// The joined diff text as it stood when the model verifier last answered
+    /// a fresh verdict on this candidate — the delta-framing baseline (#1431
+    /// remaining scope). The next escalated round renders file sections
+    /// byte-identical to this as stat lines instead of re-buying their
+    /// bodies. `None` until a model verdict has been bought, and never set by
+    /// a heuristic fallback: a verdict no model read must not let the next
+    /// round claim its diff was already reviewed.
+    last_verdict_diff: Option<String>,
 }
 
 impl CandidateState {
@@ -2136,6 +2145,7 @@ impl<'a> Pipeline<'a> {
             witness_paths: Vec::new(),
             failures: Vec::new(),
             last_verdict: None,
+            last_verdict_diff: None,
         };
 
         if let Err(reason) = self
@@ -2654,44 +2664,22 @@ impl<'a> Pipeline<'a> {
                     // a guidance call on the FIRST deterministic red while
                     // telling the verifier the agent had "failed twice in a row".
                     let mut reason = brief.message();
-                    if self.config.distress_guidance && state.failures.len() >= 2 {
-                        // Same witness exclusion as the verdict call (#1433):
-                        // guidance reads the change under correction, and the
-                        // verifier's own test is not part of it.
-                        let stripped =
-                            crate::verify::strip_witness_hunks(&state.diff_text, &witness_paths);
-                        match self
-                            .verifier_guidance(
+                    if self.config.distress_guidance
+                        && state.failures.len() >= 2
+                        && let Err(abort) = self
+                            .distress_guidance_reason(
                                 goal,
-                                &stripped.diff,
+                                &state.diff_text,
+                                &witness_paths,
                                 &evidence.summary,
-                                // Guidance never carries a delta baseline: its
-                                // render is already evidence-scoped (#1432),
-                                // and "unchanged since a verdict round" is a
-                                // verdict-shaped claim.
-                                &DiffContext {
-                                    witness_paths: &witness_paths,
-                                    previous: None,
-                                },
+                                &sealed,
+                                &mut reason,
                                 budget,
                                 total,
                             )
                             .await
-                        {
-                            Ok(Some(guidance)) => {
-                                if let Some(text) =
-                                    self.airlock_forward(&guidance, "distress_guidance", &sealed)
-                                {
-                                    reason
-                                        .push_str("\n\nIndependent reviewer course-correction:\n");
-                                    reason.push_str(&text);
-                                }
-                            }
-                            Ok(None) => {}
-                            Err(abort) => {
-                                return CandidateResult::aborted(state.messages, abort.reason);
-                            }
-                        }
+                    {
+                        return CandidateResult::aborted(state.messages, abort.reason);
                     }
                     if let Err(reason) = self
                         .revise_candidate(engine, surface, budget, &reason, total, &mut state)
@@ -2735,167 +2723,39 @@ impl<'a> Pipeline<'a> {
                     }
                     // Inconclusive — escalate to the model verifier (verifier ≠
                     // worker; a verifier-call failure falls back to a heuristic).
-                    let mut evidence_summary = format!(
-                        "flip_achieved={}; touched_tests={:?}; diff_lines={} (budget {}); \
-                         file_change_events={}",
-                        inputs.flip_achieved,
-                        inputs.touched_tests_passed,
-                        state.diff_lines,
-                        self.config.diff_budget_lines,
-                        state.file_changes,
+                    // The evidence summary assembles every deterministic
+                    // signal the ladder collected (#859, #860, #861, #864,
+                    // #867, #870, #1291) — split out alongside
+                    // `ladder_snapshot` in `verify_probes.rs`.
+                    let mut evidence_summary = self.model_verdict_evidence_summary(
+                        &inputs,
+                        &state,
+                        test_infra,
+                        &lint_sample,
+                        &snapshot,
                     );
-                    if let Some(label) = test_infra {
-                        // #860: the run ended without observing an assertion.
-                        // Named so the verifier reads "the suite timed out", not
-                        // "the suite failed".
-                        evidence_summary
-                            .push_str(&format!("; test_run={label} (no assertion observed)"));
-                    }
-                    if state.oracle.is_unstable() {
-                        // #859: a fail→pass flip WAS observed but could not
-                        // be reproduced on the same tree — a different fact
-                        // from "the test never passed", and one the verifier
-                        // should weigh explicitly.
-                        evidence_summary.push_str(
-                            "; unstable_flip=true (the flip's confirmation re-run did not pass)",
-                        );
-                    }
-                    if inputs.diff_coverage != DiffCoverage::Unmeasured {
-                        // #1291: stated only when it was actually measured.
-                        // An `unmeasured` line here would be pure noise in a
-                        // verifier prompt — the ladder already escalated for some
-                        // other reason, and "nobody looked" adds nothing to
-                        // reason from. It stays on the snapshot either way.
-                        evidence_summary.push_str("; ");
-                        evidence_summary.push_str(inputs.diff_coverage.explain());
-                    }
-                    if state.witness_mutation == Some(false) {
-                        // #870: the witness reacted to the change without
-                        // constraining it — it stayed green while the
-                        // changed lines were deliberately broken.
-                        evidence_summary.push_str(
-                            "; witness_tautological=true (the witness stayed green under every \
-                             trivial mutation of the changed lines)",
-                        );
-                    }
-                    if state.oracle.refused_different_failure() {
-                        // #867: the suite passed, but its own complete test
-                        // listing did not contain the baseline's failing
-                        // tests — the observed failure was not fixed, it
-                        // disappeared. The verifier should treat the passing
-                        // exit code accordingly.
-                        evidence_summary.push_str(
-                            "; flip_refused=the passing run's test listing does not contain \
-                             the test(s) that failed on the baseline (fixed a different \
-                             failure, or the failing test was removed)",
-                        );
-                    }
-                    if inputs.new_diag_errors > 0 || inputs.new_diag_warnings > 0 {
-                        // #861: the regression the veto saw, capped to a
-                        // 3-line sample so the verifier reads the delta, not
-                        // the linter's whole opinion.
-                        evidence_summary.push_str(&format!(
-                            "; new_diagnostics={} error(s), {} warning(s) vs baseline",
-                            inputs.new_diag_errors, inputs.new_diag_warnings
-                        ));
-                        if !lint_sample.is_empty() {
-                            evidence_summary.push('\n');
-                            evidence_summary.push_str(lint_sample.trim_end());
+                    // Witness exclusion (#1433), verdict reuse (#1431 step 1),
+                    // and the delta-framing baseline (#1431 remaining scope)
+                    // all live in `resolve_verdict` — one nameable concern,
+                    // split out for the same reason the rest of
+                    // `verifier_stage.rs` is.
+                    let verdict = match self
+                        .resolve_verdict(
+                            goal,
+                            &mut evidence_summary,
+                            &mut state,
+                            &witness_paths,
+                            &inputs,
+                            budget,
+                            total,
+                        )
+                        .await
+                    {
+                        Ok(verdict) => verdict,
+                        Err(abort) => {
+                            return CandidateResult::aborted(state.messages, abort.reason);
                         }
-                    }
-                    if !snapshot.oracle_trace.is_empty() {
-                        // #864: the oracle trace, rendered compactly. The
-                        // verifier sees WHY the ladder was inconclusive — which
-                        // runs happened, in order, and what each observed —
-                        // instead of a diff cold.
-                        evidence_summary.push_str(&format!(
-                            "; oracle_trace=[{}]",
-                            crate::replay::render_oracle_trace(&snapshot.oracle_trace)
-                        ));
-                    }
-                    if snapshot.witness_intact == Some(true) {
-                        // #864: the tamper-exclusion result, stated. A
-                        // tampered witness never reaches a verifier, so what the
-                        // verifier learns here is that the check RAN and the
-                        // witness it is weighing is the authored one.
-                        evidence_summary.push_str("; witness_tamper_check=intact");
-                    }
-                    // The witness's own chunks never ride into the paid prompt
-                    // as "worker-authored data" (#1433): they are the
-                    // verifier's own artifact, and everything the verdict needs
-                    // to know about the witness is already in the trusted
-                    // evidence above. The omission is named HERE — the trusted
-                    // zone — never in-band in the diff, where the framing says
-                    // every byte is forgeable worker data.
-                    let stripped =
-                        crate::verify::strip_witness_hunks(&state.diff_text, &witness_paths);
-                    if !stripped.omitted.is_empty() {
-                        evidence_summary.push_str(&format!(
-                            "; witness_files_omitted_from_diff=[{}] (verifier-authored test, \
-                             not part of the change under review)",
-                            stripped.omitted.join(", ")
-                        ));
-                    }
-                    // Reuse before re-buying (#1431): a revision that changed
-                    // nothing the verdict depends on — same goal, same diff,
-                    // same evidence, byte for byte — would re-ask the same
-                    // question and pay full price for sampling noise. The
-                    // cached verdict is the same opinion at zero cost, and the
-                    // appended note steers the worker better than a fresh
-                    // reading of an unchanged tree ever did.
-                    let inputs_digest =
-                        verdict_inputs_digest(goal, &stripped.diff, &evidence_summary);
-                    let cached = state
-                        .last_verdict
-                        .as_ref()
-                        .filter(|(digest, _)| *digest == inputs_digest)
-                        .map(|(_, verdict)| verdict.clone());
-                    let verdict = match cached {
-                        Some(mut verdict) => {
-                            verdict.reasoning.push_str(
-                                "\n(verdict reused: the goal, diff, and evidence are unchanged \
-                                 since the previous review round — no new model call was made)",
-                            );
-                            verdict
-                        }
-                        None => match self
-                            .verifier(
-                                goal,
-                                &stripped.diff,
-                                &evidence_summary,
-                                &inputs,
-                                budget,
-                                total,
-                            )
-                            .await
-                        {
-                            Ok(verdict) => {
-                                // Only pin a real model verdict for reuse. A
-                                // heuristic fallback is a transient-outage
-                                // stand-in (unresolvable provider, unparseable
-                                // response, failed/timed-out call), not the
-                                // opinion this candidate bought: caching it
-                                // would suppress recovery on the next round
-                                // (the verifier may have come back) and graft
-                                // the "no new model call was made" reuse note
-                                // onto a fallback that never made one.
-                                if !verdict.heuristic {
-                                    state.last_verdict = Some((inputs_digest, verdict.clone()));
-                                }
-                                verdict
-                            }
-                            Err(abort) => {
-                                return CandidateResult::aborted(state.messages, abort.reason);
-                            }
-                        },
                     };
-                    if !verdict.heuristic {
-                        // The delta-framing baseline (#1431) advances only on
-                        // a verdict a model actually answered: a heuristic
-                        // fallback read nothing, and must not let the next
-                        // round stat-line text no model ever saw.
-                        state.last_verdict_diff = Some(state.diff_text.clone());
-                    }
                     let mut evidence = model_verdict_evidence(&verdict);
                     evidence.ladder = Some(Box::new(snapshot.with_rung(verdict.rung())));
                     self.emit(AgentEvent::Verdict {
@@ -3810,20 +3670,6 @@ fn assemble_user_message(
         VerificationContract::None => {}
     }
     s
-}
-
-/// One digest over everything a model verdict depends on — goal, the (witness
-/// -stripped) diff, and the evidence summary. Byte-identical inputs are the
-/// same question; the ModelVerdict arm reuses its previous answer rather than
-/// paying for sampling noise (#1431). Within-run only, so the std hasher's
-/// stability across versions is irrelevant.
-fn verdict_inputs_digest(goal: &str, diff: &str, evidence_summary: &str) -> u64 {
-    use std::hash::{Hash, Hasher};
-    let mut hasher = std::collections::hash_map::DefaultHasher::new();
-    goal.hash(&mut hasher);
-    diff.hash(&mut hasher);
-    evidence_summary.hash(&mut hasher);
-    hasher.finish()
 }
 
 /// The instruction appended to a revision turn, carrying the failing
