+++ crates/stella-pipeline/src/pipeline/tests.rs
@@ -4,7 +4,10 @@
 //! stays reachable via `super::*`.
 
 mod management_accounting;
+mod scripted_provider;
 mod telemetry;
+
+use scripted_provider::ScriptedProvider;
 
 use super::*;
 use crate::CmdKind;
@@ -201,65 +204,6 @@ impl RepoStatusPort for SeqRepoStatus {
                     link_count: 1,
                 })
         })
-    }
-}
-
-/// A scripted provider serving triage, plan, worker, and verifier calls from
-/// one ordered queue (the resolver hands the same provider to every role).
-struct ScriptedProvider {
-    script: TokioMutex<VecDeque<CompletionResult>>,
-    /// Every prompt this provider was asked to complete, in order — so a test
-    /// can assert what actually reached the model, not just what came back.
-    seen: std::sync::Mutex<Vec<String>>,
-    /// The same requests with per-message roles preserved, so a test can
-    /// assert the system/user split of a management call (#1434) — the joined
-    /// text above cannot show which half a sentence landed in.
-    shapes: std::sync::Mutex<Vec<Vec<(stella_protocol::MessageRole, String)>>>,
-}
-impl ScriptedProvider {
-    fn new(results: Vec<CompletionResult>) -> Self {
-        Self {
-            script: TokioMutex::new(results.into_iter().collect()),
-            seen: std::sync::Mutex::new(Vec::new()),
-            shapes: std::sync::Mutex::new(Vec::new()),
-        }
-    }
-
-    /// The message text of each request, one joined string per call.
-    fn prompts(&self) -> Vec<String> {
-        self.seen.lock().unwrap().clone()
-    }
-
-    /// Each request's messages as `(role, content)` pairs, in call order.
-    fn shapes(&self) -> Vec<Vec<(stella_protocol::MessageRole, String)>> {
-        self.shapes.lock().unwrap().clone()
-    }
-}
-#[async_trait]
-impl Provider for ScriptedProvider {
-    fn id(&self) -> &str {
-        "scripted"
-    }
-    async fn complete_ref(
-        &self,
-        req: CompletionRequestRef<'_>,
-    ) -> Result<CompletionResult, ProviderError> {
-        self.seen.lock().unwrap().push(
-            req.messages
-                .iter()
-                .map(|m| m.content.as_str())
-                .collect::<Vec<_>>()
-                .join("\n"),
-        );
-        self.shapes.lock().unwrap().push(
-            req.messages
-                .iter()
-                .map(|m| (m.role, m.content.clone()))
-                .collect(),
-        );
-        let mut q = self.script.lock().await;
-        q.pop_front()
-            .ok_or_else(|| ProviderError::Terminal("scripted provider exhausted".into()))
     }
 }
