+++ crates/stella-pipeline/src/verify.rs
@@ -786,17 +786,6 @@ pub fn model_verdict_evidence(verdict: &Verdict) -> VerdictEvidence {
     }
 }
 
-/// Ceiling on the worker-authored diff text interpolated into a verifier or
-/// guidance prompt, in chars (~10k tokens). The fast-submit diff budget is
-/// 400 *lines*, so a legitimately judged diff almost always fits whole; what
-/// this bounds is the pathological tail — a generated file, a vendored blob, a
-/// worker that rewrote the world — which used to ride into every paid verifier
-/// call at full length. Head-weighted middle-out, matching the compactor's
-/// aging pass: the head carries the file headers and the intent, the tail the
-/// most recent hunks, and the elision is marked in-band so the verifier knows it
-/// is reading an excerpt rather than the whole change.
-const VERIFIER_DIFF_BUDGET_CHARS: usize = 40_000;
-
 /// A diff with the authored witness's own file chunks removed, plus the
 /// paths that were removed — see [`strip_witness_hunks`].
 pub struct StrippedDiff {
@@ -888,26 +877,21 @@ pub fn strip_witness_hunks(diff: &str, witness_paths: &[String]) -> StrippedDiff
     StrippedDiff { diff: out, omitted }
 }
 
-/// Clamp a worker-authored diff to `VERIFIER_DIFF_BUDGET_CHARS` for prompt
-/// interpolation: keep the head and tail, elide the middle, and say so where
-/// the cut was made. Char-based, not byte-based, so a multi-byte diff can
-/// never split a code point (the same unit [`crate::pipeline`]'s recall
-/// clamp settled on).
-fn bounded_worker_diff(diff: &str) -> String {
-    let total = diff.chars().count();
-    if total <= VERIFIER_DIFF_BUDGET_CHARS {
-        return diff.to_string();
-    }
-    let head_chars = VERIFIER_DIFF_BUDGET_CHARS * 2 / 3;
-    let tail_chars = VERIFIER_DIFF_BUDGET_CHARS - head_chars;
-    let head: String = diff.chars().take(head_chars).collect();
-    let tail: String = diff.chars().skip(total - tail_chars).collect();
-    let elided = total - head_chars - tail_chars;
-    format!(
-        "{head}\n[… {elided} chars elided from the middle of the diff — the head and tail are \
-         shown; verifier from what is visible and weigh that the middle is not …]\n{tail}"
-    )
-}
+/// The fixed sentence both verifier-facing prompts carry to explain the
+/// renderer's `#` stat lines ([`diff_render`]). One constant, for the same
+/// reason [`UNTRUSTED_DIFF_HEADING_SUFFIX`] is one: prompts and tests must
+/// read the same spelling or the guard outlives the thing it guards.
+///
+/// Unconditional — present even when the render happened to reduce nothing —
+/// because the instruction text is exactly the part of these prompts that
+/// must stay byte-stable across calls (the management-call caching work,
+/// #1434, depends on that stability).
+const DIFF_STAT_LINE_NOTE: &str = "Inside the diff, a line beginning with `#` is a rendering note from the pipeline, not \
+     part of the change: a file section may be reduced to one such stat line when it is \
+     unchanged since a previous review round of this same candidate (a prior round read its \
+     full text), when it is the pipeline's own witness test rather than the worker's change, \
+     or when the diff exceeds its token budget. A summarized file is still part of the \
+     change — weigh what its stat line states.";
 
 /// The one framing under which worker-authored text may enter a verifier-facing
 /// prompt (witness-protocol D5, `docs/design/witness-protocol.md` §2): the
